+++ src/main/java/com/fein91/dao/OrderRequestRepository.java
@@ -10,12 +10,12 @@ import java.math.BigInteger;
 import java.util.List;
 
 /**
- * Created by olta1014 on 23.05.2016.
+ * Order request repository
  */
-public interface OrderRequestRepository extends CrudRepository<OrderRequest, BigInteger> {
+public interface OrderRequestRepository extends CrudRepository<OrderRequest, Long> {
 
     List<OrderRequest> findByCounterpartyAndOrderSide(Counterparty counterparty, int orderSide);
 
     @Query("SELECT r FROM OrderRequest r where r.counterparty.id = :counterPartyId")
-    List<OrderRequest> findByCounterpartyId(@Param("counterPartyId") BigInteger counterPartyId);
+    List<OrderRequest> findByCounterpartyId(@Param("counterPartyId") Long counterPartyId);
 }
